export MultiValuePill into separate component
